+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v0.1.3-r31 (2026-08-29)
+
+### Added
+
+- Added `--min_freq` and `--max_occ` to filter frequent minimizers and speed up alignment.
+- Added `--min_ali_len` to filter short, potentially false alignments.
+
 ## v0.1.3-r30 (2026-08-28)
 
 ### Added